Cleaned code lines, updated part2 with a practice session, broke up exercises into part 1 &2
All code cells are now without line numbers which is much cleaner.
Part now has a practice session like part 1 where students don't have to type out the contents but are asked to workout what they expect the output to be. The exercises are now seperated into exercises for part1 and exercises for part2. I've also added an initial explanation of the format used (i.e. docstrings). I've also added an extra cell in the strings section of part1 showing how to construct raw strings (as they are used in the docstrings)